chore(repo): move skin and chat package to board (#305)
* move skin and chat package to board

* fix
