+++ packages/board/package.json
@@ -13,9 +13,8 @@
     "@types/react-dom": "^17.0.11"
   },
   "dependencies": {
-    "@botpress/chat": "0.0.1",
-    "@botpress/chat-skin": "0.0.1",
     "@botpress/messaging-client": "*",
+    "@botpress/messaging-socket": "0.0.2",
     "@botpress/webchat": "0.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
