+++ packages/board/src/chat/base/system.ts
