+++ packages/board/src/chat/chat.ts
