+++ packages/board/src/chat/conversation/events.ts
