+++ packages/board/src/chat/index.ts
