+++ packages/board/src/chat/locale/system.ts
