+++ packages/board/src/chat/messages/system.ts
