+++ packages/board/src/skin/lang/en.json
