+++ packages/board/src/skin/lang/fr.json
