+++ packages/board/src/skin/lang/index.ts
