--- packages/chat/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@botpress/chat",
-  "version": "0.0.1",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "source": "src/index.ts",
-  "license": "AGPL-3.0",
-  "scripts": {
-    "build": "yarn --silent && yarn run -T tsc --build",
-    "watch": "yarn --silent && yarn run -T tsc --build --watch",
-    "prepublishOnly": "yarn --silent && yarn run -T tsc --build"
-  },
-  "files": [
-    "dist"
-  ],
-  "dependencies": {
-    "@botpress/messaging-base": "0.0.1",
-    "@botpress/messaging-socket": "0.0.2"
-  }
-}
