--- packages/chat/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  // Manage references here
-  "references": [
-    {
-      "path": "../socket"
-    }
-  ],
-
-  // Defaults (don't change this)
-  "extends": "../../tsconfig.packages.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist",
-    "tsBuildInfoFile": "dist/.tsbuildinfo",
-
-    // Specific settings for npm package
-    "target": "es6",
-    "declaration": true,
-    "sourceMap": false,
-    "lib": ["dom", "es2021"]
-  },
-  "exclude": ["dist", "node_modules"]
-}
