--- packages/skin/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@botpress/chat-skin",
-  "version": "0.0.1",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "source": "src/index.ts",
-  "license": "AGPL-3.0",
-  "scripts": {
-    "build": "yarn --silent && yarn run -T tsc --build",
-    "watch": "yarn --silent && yarn run -T tsc --build --watch"
-  },
-  "files": [
-    "dist"
-  ],
-  "dependencies": {
-    "@botpress/chat": "0.0.1",
-    "@botpress/messaging-base": "0.0.1"
-  }
-}
