--- packages/skin/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  // Manage references here
-  "references": [
-    {
-      "path": "../chat"
-    }
-  ],
-
-  // Defaults (don't change this)
-  "extends": "../../tsconfig.packages.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist",
-    "tsBuildInfoFile": "dist/.tsbuildinfo",
-
-    // Specific settings for npm package
-    "target": "es6",
-    "declaration": true,
-    "sourceMap": false,
-    "resolveJsonModule": true
-  },
-  "include": ["src", "src/**/*.json"],
-  "exclude": ["dist", "node_modules"]
-}
